==> 1-train-test-split.ipynb <==
import pandas as pd
import numpy as np
import ast
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MultiLabelBinarizer

file_path = '/kaggle/input/movie-dataset-filtered/movies_metadata_filtered.csv'
data = pd.read_csv(file_path, low_memory=False)
data['genres'] = data['transformed_genres'].apply(ast.literal_eval)
print(data.shape)
print(data.head())

subset = data.copy()

mlb = MultiLabelBinarizer()
subset = subset.join(pd.DataFrame(mlb.fit_transform(subset.pop('genres')),
                                  columns=mlb.classes_,
                                  index=subset.index))

train, temp = train_test_split(subset, test_size=0.2, random_state=23)
val, test = train_test_split(temp, test_size=0.5, random_state=23)

train_path = '/kaggle/working/train.csv'
val_path = '/kaggle/working/val.csv'
test_path = '/kaggle/working/test.csv'

train.to_csv(train_path, index=False)
val.to_csv(val_path, index=False)
test.to_csv(test_path, index=False)

print(f"Train set saved to {train_path}")
print(f"Validation set saved to {val_path}")
print(f"Test set saved to {test_path}")

train.head()
